+++ unit_tests/reduct/api/entry_api_test.cc
@@ -42,63 +42,44 @@ TEST_CASE("EntryApi::Write should write data in chunks") {
     }
 
     SECTION("record already exists") {
-      auto [receiver, err] = EntryApi::Write(storage.get(), "bucket", "entry-1", "1000001", "10");
-      REQUIRE(err == Error::kOk);
-      REQUIRE(receiver("", true).error == Error{
-                                              .code = 409,
-                                              .message = "A record with timestamp 1000001 already exists",
-                                          });
+      REQUIRE(EntryApi::Write(storage.get(), "bucket", "entry-1", "1000001", "10").error ==
+              Error::Conflict("A record with timestamp 1000001 already exists"));
     }
   }
 
   SECTION("wrong ts") {
-    auto [receiver, err] = EntryApi::Write(storage.get(), "bucket", "entry-1", "XXXX", "10");
-    REQUIRE(err == Error::kOk);
-    REQUIRE(receiver("", true).error ==
-            Error{
-                .code = 422,
-                .message = "Failed to parse 'ts' parameter: XXXX must unix times in microseconds",
-            });
+    REQUIRE(EntryApi::Write(storage.get(), "bucket", "entry-1", "XXXX", "10").error ==
+            Error::UnprocessableEntity("Failed to parse 'ts' parameter: XXXX must unix times in microseconds"));
   }
 
   SECTION("negative ts") {
-    auto [receiver, err] = EntryApi::Write(storage.get(), "bucket", "entry-1", "-100", "10");
-    REQUIRE(err == Error::kOk);
-    REQUIRE(receiver("", true).error == Error{
-                                            .code = 422,
-                                            .message = "Failed to parse 'ts' parameter: -100 must be positive",
-                                        });
+    REQUIRE(EntryApi::Write(storage.get(), "bucket", "entry-1", "-100", "10").error ==
+            Error::UnprocessableEntity("Failed to parse 'ts' parameter: -100 must be positive"));
   }
 
   SECTION("empty entry") {
-    auto [receiver, err] = EntryApi::Write(storage.get(), "bucket", "", "1000001", "10");
-    REQUIRE(err == Error::kOk);
-    REQUIRE(receiver("", true).error == Error{
-                                            .code = 422,
-                                            .message = "An empty entry name is not allowed",
-                                        });
+    REQUIRE(EntryApi::Write(storage.get(), "bucket", "", "1000001", "10").error ==
+            Error::UnprocessableEntity("An empty entry name is not allowed"));
   }
 
   SECTION("wrong content-length") {
     auto entry = storage->GetBucket("bucket").result.lock()->GetOrCreateEntry("entry-1").result.lock();
 
     SECTION("empty") {
-      auto [receiver, err] = EntryApi::Write(storage.get(), "bucket", "entry-1", "1000001", "");
-      REQUIRE(err == Error::kOk);
-      REQUIRE(receiver("", true).error == Error::ContentLengthRequired("Bad or empty content-length"));
+      REQUIRE(EntryApi::Write(storage.get(), "bucket", "entry-1", "1000001", "").error ==
+              Error::ContentLengthRequired("Bad or empty content-length"));
     }
 
     SECTION("negative") {
-      auto [receiver, err] = EntryApi::Write(storage.get(), "bucket", "entry-1", "1000001", "-1");
-      REQUIRE(err == Error::kOk);
-      REQUIRE(receiver("", true).error == Error::ContentLengthRequired("Negative content-length"));
+      REQUIRE(EntryApi::Write(storage.get(), "bucket", "entry-1", "1000001", "-1").error ==
+              Error::ContentLengthRequired("Negative content-length"));
     }
 
     SECTION("no number") {
-      auto [receiver, err] = EntryApi::Write(storage.get(), "bucket", "entry-1", "1000001", "xxxx");
-      REQUIRE(err == Error::kOk);
-      REQUIRE(receiver("", true).error == Error::ContentLengthRequired("Bad or empty content-length"));
+      REQUIRE(EntryApi::Write(storage.get(), "bucket", "entry-1", "1000001", "xxxx").error ==
+              Error::ContentLengthRequired("Bad or empty content-length"));
     }
+
     REQUIRE(entry->BeginRead(reduct::core::Time() + us(1000001)).error.code == 404);
   }
 
@@ -120,14 +101,8 @@ TEST_CASE("EntryApi::Write should write data in chunks") {
     }
 
     SECTION("too short") {
-      REQUIRE(receiver("1234", true).error == Error{
-                                                  .code = 413,
-                                                  .message = "Content is smaller than in content-length",
-                                              });
-      REQUIRE(entry->BeginRead(reduct::core::Time() + us(1000001)).error == Error{
-                                                                                .code = 500,
-                                                                                .message = "Record is broken",
-                                                                            });
+      REQUIRE(receiver("1234", true).error == Error(413, "Content is smaller than in content-length"));
+      REQUIRE(entry->BeginRead(reduct::core::Time() + us(1000001)).error == Error::InternalError("Record is broken"));
     }
   }
 }
@@ -157,42 +132,23 @@ TEST_CASE("EntryApi::Read should read data in chunks with time") {
   }
 
   SECTION("bucket doesn't exist") {
-    auto [receiver, err] = EntryApi::Read(storage.get(), "XXX", "entry-1", "1000001", {});
-    REQUIRE(err == Error::kOk);
-
-    REQUIRE(receiver("", true).error == Error{
-                                            .code = 404,
-                                            .message = "Bucket 'XXX' is not found",
-                                        });
+    REQUIRE(EntryApi::Read(storage.get(), "XXX", "entry-1", "1000001", {}).error ==
+            Error::NotFound("Bucket 'XXX' is not found"));
   }
 
   SECTION("entry doesn't exist") {
-    auto [receiver, err] = EntryApi::Read(storage.get(), "bucket", "XXX", "1000001", {});
-    REQUIRE(err == Error::kOk);
-
-    REQUIRE(receiver("", true).error == Error{
-                                            .code = 404,
-                                            .message = "Entry 'XXX' is not found",
-                                        });
+    REQUIRE(EntryApi::Read(storage.get(), "bucket", "XXX", "1000001", {}).error ==
+            Error::NotFound("Entry 'XXX' is not found"));
   }
 
   SECTION("record doesn't exist") {
-    auto [receiver, err] = EntryApi::Read(storage.get(), "bucket", "entry-1", "1234567", {});
-    REQUIRE(err == Error::kOk);
-    REQUIRE(receiver("", true).error == Error{
-                                            .code = 404,
-                                            .message = "No records for this timestamp",
-                                        });
+    REQUIRE(EntryApi::Read(storage.get(), "bucket", "entry-1", "1234567", {}).error ==
+            Error::NotFound("No records for this timestamp"));
   }
 
   SECTION("wrong ts") {
-    auto [receiver, err] = EntryApi::Write(storage.get(), "bucket", "entry-1", "XXXX", {});
-    REQUIRE(err == Error::kOk);
-    REQUIRE(receiver("", true).error ==
-            Error{
-                .code = 422,
-                .message = "Failed to parse 'ts' parameter: XXXX must unix times in microseconds",
-            });
+    REQUIRE(EntryApi::Write(storage.get(), "bucket", "entry-1", "XXXX", {}).error ==
+            Error::UnprocessableEntity("Failed to parse 'ts' parameter: XXXX must unix times in microseconds"));
   }
 }
 
@@ -239,31 +195,18 @@ TEST_CASE("EntryApi::Read should read data in chunks with query id") {
       REQUIRE(ret.result == "abcd");
     }
 
-    {
-      auto [receiver, err] = EntryApi::Read(storage.get(), "bucket", "entry-1", {}, id);
-      REQUIRE(err == Error::kOk);
-
-      REQUIRE(receiver("", true).error ==
-              Error{.code = 404, .message = fmt::format("Query id={} doesn't exist. It expired or was finished", id)});
-    }
+    REQUIRE(EntryApi::Read(storage.get(), "bucket", "entry-1", {}, id).error ==
+            Error::NotFound(fmt::format("Query id={} doesn't exist. It expired or was finished", id)));
   }
 
   SECTION("id doesnt' exist") {
-    auto [receiver, err] = EntryApi::Read(storage.get(), "bucket", "entry-1", {}, "2");
-    REQUIRE(err == Error::kOk);
-    REQUIRE(receiver("", true).error == Error{
-                                            .code = 404,
-                                            .message = "Query id=2 doesn't exist. It expired or was finished",
-                                        });
+    REQUIRE(EntryApi::Read(storage.get(), "bucket", "entry-1", {}, "2").error ==
+            Error::NotFound("Query id=2 doesn't exist. It expired or was finished"));
   }
 
   SECTION("wrong id") {
-    auto [receiver, err] = EntryApi::Read(storage.get(), "bucket", "entry-1", {}, "XXX2");
-    REQUIRE(err == Error::kOk);
-    REQUIRE(receiver("", true).error == Error{
-                                            .code = 422,
-                                            .message = "Failed to parse 'id' parameter: XXX2 must be unsigned integer",
-                                        });
+    REQUIRE(EntryApi::Read(storage.get(), "bucket", "entry-1", {}, "XXX2").error ==
+            Error::UnprocessableEntity("Failed to parse 'id' parameter: XXX2 must be unsigned integer"));
   }
 }
 
@@ -289,7 +232,7 @@ TEST_CASE("EntryApi::Query should query data for time interval") {
 
     REQUIRE(entry->Next(info.id()).error == Error::kOk);
     REQUIRE(entry->Next(info.id()).error == Error::kOk);
-    REQUIRE(entry->Next(info.id()).error.code == 404);
+    REQUIRE(entry->Next(info.id()).error.code == Error::kNotFound);
   }
 
   SECTION("ok start") {
@@ -302,7 +245,7 @@ TEST_CASE("EntryApi::Query should query data for time interval") {
     JsonStringToMessage(resp.SendData().result, &info);
 
     REQUIRE(entry->Next(info.id()).result.reader->timestamp() == Time() + us(2000001));
-    REQUIRE(entry->Next(info.id()).error.code == 404);
+    REQUIRE(entry->Next(info.id()).error.code == Error::kNotFound);
   }
 
   SECTION("ok stop") {
@@ -315,7 +258,7 @@ TEST_CASE("EntryApi::Query should query data for time interval") {
     JsonStringToMessage(resp.SendData().result, &info);
 
     REQUIRE(entry->Next(info.id()).result.reader->timestamp() == Time() + us(1000001));
-    REQUIRE(entry->Next(info.id()).error.code == 404);
+    REQUIRE(entry->Next(info.id()).error.code == Error::kNotFound);
   }
 
   SECTION("ok timeinterval") {
@@ -327,8 +270,8 @@ TEST_CASE("EntryApi::Query should query data for time interval") {
 
     JsonStringToMessage(resp.SendData().result, &info);
 
-    REQUIRE(entry->Next(info.id()).error.code == 202);  // no content
-    REQUIRE(entry->Next(info.id()).error.code == 404);
+    REQUIRE(entry->Next(info.id()).error.code == Error::kNoContent);  // no content
+    REQUIRE(entry->Next(info.id()).error.code == Error::kNotFound);
   }
 
   SECTION("ok ttl") {
@@ -336,63 +279,29 @@ TEST_CASE("EntryApi::Query should query data for time interval") {
     REQUIRE(err == Error::kOk);
 
     std::this_thread::sleep_for(us(1'000'000));
-    REQUIRE(entry->Next(info.id()).error.code == 404);
+    REQUIRE(entry->Next(info.id()).error.code == Error::kNotFound);
   }
 
   SECTION("bucket doesn't exist") {
-    auto [receiver, err] = EntryApi::Query(storage.get(), "XXX", "entry-1", {}, {}, {});
-    REQUIRE(err == Error::kOk);
-    REQUIRE(receiver("", true).error == Error{
-                                            .code = 404,
-                                            .message = "Bucket 'XXX' is not found",
-                                        });
+    REQUIRE(EntryApi::Query(storage.get(), "XXX", "entry-1", {}, {}, {}).error ==
+            Error::NotFound("Bucket 'XXX' is not found"));
   }
 
   SECTION("entry doesn't exist") {
-    auto [receiver, err] = EntryApi::Query(storage.get(), "bucket", "XXX", {}, {}, {});
-    REQUIRE(err == Error::kOk);
-    REQUIRE(receiver("", true).error == Error{
-                                            .code = 404,
-                                            .message = "Entry 'XXX' is not found",
-                                        });
+    REQUIRE(EntryApi::Query(storage.get(), "bucket", "XXX", {}, {}, {}).error ==
+            Error::NotFound("Entry 'XXX' is not found"));
   }
 
   SECTION("wrong parameters") {
-    {
-      auto [receiver, err] = EntryApi::Query(storage.get(), "bucket", "entry-1", "XXX", {}, {});
-      REQUIRE(err == Error::kOk);
-      REQUIRE(receiver("", true).error ==
-              Error{
-                  .code = 422,
-                  .message = "Failed to parse 'start_timestamp' parameter: XXX must unix times in microseconds",
-              });
-    }
-    {
-      auto [receiver, err] = EntryApi::Query(storage.get(), "bucket", "entry-1", {}, "XXX", {});
-      REQUIRE(err == Error::kOk);
-      REQUIRE(receiver("", true).error ==
-              Error{
-                  .code = 422,
-                  .message = "Failed to parse 'stop_timestamp' parameter: XXX must unix times in microseconds",
-              });
-    }
-    {
-      auto [receiver, err] = EntryApi::Query(storage.get(), "bucket", "entry-1", {}, {}, "XXX");
-      REQUIRE(err == Error::kOk);
-      REQUIRE(receiver("", true).error ==
-              Error{
-                  .code = 422,
-                  .message = "Failed to parse 'ttl' parameter: XXX must be unsigned integer",
-              });
-    }
-    {
-      auto [receiver, err] = EntryApi::Query(storage.get(), "bucket", "entry-1", {}, {}, "XXX");
-      REQUIRE(err == Error::kOk);
-      REQUIRE(receiver("", true).error ==
-              Error{
-                  .code = 422,
-                  .message = "Failed to parse 'ttl' parameter: XXX must be unsigned integer",
-              });
-    }
+    REQUIRE(
+        EntryApi::Query(storage.get(), "bucket", "entry-1", "XXX", {}, {}).error ==
+        Error::UnprocessableEntity("Failed to parse 'start_timestamp' parameter: XXX must unix times in microseconds"));
+    REQUIRE(
+        EntryApi::Query(storage.get(), "bucket", "entry-1", {}, "XXX", {}).error ==
+        Error::UnprocessableEntity("Failed to parse 'stop_timestamp' parameter: XXX must unix times in microseconds"));
+    REQUIRE(EntryApi::Query(storage.get(), "bucket", "entry-1", {}, {}, "XXX").error ==
+            Error::UnprocessableEntity("Failed to parse 'ttl' parameter: XXX must be unsigned integer"));
+    REQUIRE(EntryApi::Query(storage.get(), "bucket", "entry-1", {}, {}, "XXX").error ==
+            Error::UnprocessableEntity("Failed to parse 'ttl' parameter: XXX must be unsigned integer"));
   }
 }